requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-multipart
cryptography
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv
requests
